split up interface edit and rename
