+++ src/main/java/interface_adapter/edit_note/EditNotePresenter.java
@@ -38,13 +38,4 @@ public class EditNotePresenter implements EditNoteOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
-    // Method to prepare the title change view
-    @Override
-    public void prepareTitleChange(RenameNoteOutputData title) {
-        EditNoteState noteState = editNoteViewModel.getState();
-        noteState.setNoteTitle(title.getTitle());
-        this.editNoteViewModel.setState(noteState);
-        this.editNoteViewModel.firePropertyChanged();
-    }
-
 }
