+++ src/main/java/use_case/rename_note/RenameNoteOutputBoundary.java
@@ -1,0 +1,5 @@
+package use_case.rename_note;
+
+public interface RenameNoteOutputBoundary {
+    void prepareNewTitle(RenameNoteOutputData title);
+}
